analyse and show size of shared libraries
The application binary code alone is not the full attack-surface, also
analyse and include the code size of the dynamically linked libraries
that are loaded into the process.

On macOS this is more complex than one would expect as libraries are
commonly linked and loaded as framework paths. The library loader then
resolves the framework paths into the actual file path on disk. We use
dyld_info to do this work for us.